+++ restclient/src/main/java/esa/restclient/RequestBodyContent.java
@@ -10,23 +10,23 @@ public final class RequestBodyContent<T> implements BodyContent<T> {
      * Use this flag to represent content type to avoid using Class.equal()
      * to improve performance
      */
-    private final int type;
+    private final byte type;
 
     public static final class TYPE {
-        public static final int BYTES = 1;
-        public static final int FILE = 2;
-        public static final int MULTIPART = 3;
+        public static final byte BYTES = 1;
+        public static final byte FILE = 2;
+        public static final byte MULTIPART = 3;
     }
 
     private final T content;
 
-    private RequestBodyContent(int type, T content) {
+    private RequestBodyContent(byte type, T content) {
         this.type = type;
         this.content = content;
     }
 
     @Override
-    public int type() {
+    public byte type() {
         return type;
     }
